enum value: Int -> Long
GitOrigin-RevId: 49dac3875860a7da527b0694e84ce7757f1423e7

+++ gdscript/src/main/kotlin/gdscript/completion/utils/GdEnumCompletionUtil.kt
@@ -10,7 +10,7 @@ import gdscript.psi.GdEnumValue
 
 object GdEnumCompletionUtil {
 
-    fun lookup(name: String, value: Int? = null): LookupElement {
+    fun lookup(name: String, value: Long? = null): LookupElement {
         return GdLookup.create(
             name,
             icon = GdIcon.getEditorIcon(GdIcon.ENUM_MARKER),
